requirements.txt
fastapi
uvicorn
pandas
tiktoken
azure-identity
azure-search-documents
azure-core
numpy
gremlinpython
azure-cosmos
aiohttp
async_timeout
pydantic
python-dotenv
openai
fastparquet
pyarrow